// src/app/Model/departments.ts
export class departments{
    companyId: string;
    divisionId: string;
    departmentId: string;
    employeeDepartmentId: string;
    employeeDepartmentDescription: string;
    glaccountNumber: string;
    lockedBy:Date;
    lockTs: Date;
    branchCode: string;
  data: string;
  status: string;

}   